+++ extras/test/BufferingPrintTest.cpp
@@ -6,17 +6,17 @@
 
 #include "StreamUtils/Prints/BufferingPrint.hpp"
 #include "StreamUtils/Prints/SpyingPrint.hpp"
-#include "StreamUtils/Streams/MemoryStream.hpp"
+#include "StreamUtils/Prints/StringPrint.hpp"
 
 #include "doctest.h"
 
 using namespace StreamUtils;
 
 TEST_CASE("BufferingPrint") {
-  MemoryStream target(64);
+  StringPrint target;
 
-  MemoryStream actions(64);
-  SpyingPrint spy{target, actions};
+  StringPrint log;
+  SpyingPrint spy{target, log};
 
   GIVEN("capacity is 4") {
     BufferingPrint bufferingPrint{spy, 4};
@@ -26,11 +26,11 @@ TEST_CASE("BufferingPrint") {
       bufferingPrint.flush();
 
 #if STREAMUTILS_PRINT_FLUSH_EXISTS
-      CHECK(actions.readString() ==
+      CHECK(log.str() ==
             "write('ABC', 3) -> 3"
             "flush()");
 #else
-      CHECK(actions.readString() == "write('ABC', 3) -> 3");
+      CHECK(log.str() == "write('ABC', 3) -> 3");
 #endif
     }
 
@@ -39,7 +39,7 @@ TEST_CASE("BufferingPrint") {
         int n = bufferingPrint.write('A');
 
         CHECK(n == 1);
-        CHECK(actions.readString() == "");
+        CHECK(log.str() == "");
       }
 
       SUBCASE("write(uint8_t) should flush") {
@@ -49,27 +49,27 @@ TEST_CASE("BufferingPrint") {
         bufferingPrint.write('D');
         bufferingPrint.write('E');
 
-        CHECK(actions.readString() == "write('ABCD', 4) -> 4");
+        CHECK(log.str() == "write('ABCD', 4) -> 4");
       }
 
       SUBCASE("write(char*,3) goes in buffer") {
         size_t n = bufferingPrint.write("ABC", 3);
 
         CHECK(n == 3);
-        CHECK(actions.readString() == "");
+        CHECK(log.str() == "");
       }
 
       SUBCASE("write(char*,4) bypasses buffer") {
         size_t n = bufferingPrint.write("ABCD", 4);
 
         CHECK(n == 4);
-        CHECK(actions.readString() == "write('ABCD', 4) -> 4");
+        CHECK(log.str() == "write('ABCD', 4) -> 4");
       }
       SUBCASE("write(char*,2) bypasses buffer") {
         size_t n = bufferingPrint.write("ABCD", 4);
 
         CHECK(n == 4);
-        CHECK(actions.readString() == "write('ABCD', 4) -> 4");
+        CHECK(log.str() == "write('ABCD', 4) -> 4");
       }
     }
 
@@ -80,14 +80,14 @@ TEST_CASE("BufferingPrint") {
         size_t n = bufferingPrint.write("BCD", 3);
 
         CHECK(n == 3);
-        CHECK(actions.readString() == "write('ABCD', 4) -> 4");
+        CHECK(log.str() == "write('ABCD', 4) -> 4");
       }
 
       SUBCASE("write(char*,7) bypasses") {
         size_t n = bufferingPrint.write("BCDEFGH", 7);
 
         CHECK(n == 7);
-        CHECK(actions.readString() ==
+        CHECK(log.str() ==
               "write('ABCD', 4) -> 4"
               "write('EFGH', 4) -> 4");
       }
@@ -105,23 +105,23 @@ TEST_CASE("BufferingPrint") {
       int n = bufferingPrint.write('X');
 
       CHECK(n == 1);
-      CHECK(actions.readString() == "write('X') -> 1");
+      CHECK(log.str() == "write('X') -> 1");
     }
 
     SUBCASE("write(char*,1) forwards to target") {
       int n = bufferingPrint.write("A", 1);
 
       CHECK(n == 1);
-      CHECK(actions.readString() == "write('A', 1) -> 1");
+      CHECK(log.str() == "write('A', 1) -> 1");
     }
 
     SUBCASE("flush()") {
       bufferingPrint.flush();
 
 #if STREAMUTILS_PRINT_FLUSH_EXISTS
-      CHECK(actions.readString() == "flush()");
+      CHECK(log.str() == "flush()");
 #else
-      CHECK(actions.readString() == "");
+      CHECK(log.str() == "");
 #endif
     }
   }
@@ -132,6 +132,6 @@ TEST_CASE("BufferingPrint") {
       bufferingPrint.write("ABC", 3);
     }
 
-    CHECK(actions.readString() == "write('ABC', 3) -> 3");
+    CHECK(log.str() == "write('ABC', 3) -> 3");
   }
 }