+++ extras/test/ReadLoggingClientTest.cpp
@@ -7,6 +7,7 @@
 #include "StreamUtils/Clients/MemoryClient.hpp"
 #include "StreamUtils/Clients/ReadLoggingClient.hpp"
 #include "StreamUtils/Clients/SpyingClient.hpp"
+#include "StreamUtils/Prints/StringPrint.hpp"
 
 #include "doctest.h"
 
@@ -15,11 +16,11 @@ using namespace StreamUtils;
 TEST_CASE("ReadLoggingClient") {
   MemoryClient target(4);
 
-  MemoryClient actions(64);
-  SpyingClient upstreamSpy{target, actions};
+  StringPrint log;
+  SpyingClient upstreamSpy{target, log};
 
-  MemoryStream log(64);
-  ReadLoggingClient loggingClient{upstreamSpy, log};
+  StringPrint output;
+  ReadLoggingClient loggingClient{upstreamSpy, output};
 
   SUBCASE("available()") {
     target.print("ABC");
@@ -27,47 +28,47 @@ TEST_CASE("ReadLoggingClient") {
     size_t n = loggingClient.available();
 
     CHECK(n == 3);
-    CHECK(actions.readString() == "available() -> 3");
-    CHECK(log.readString() == "");
+    CHECK(log.str() == "available() -> 3");
+    CHECK(output.str() == "");
   }
 
   SUBCASE("connect(IPAddress)") {
     int n = loggingClient.connect(IPAddress("1.2.3.4"), 80);
 
     CHECK(n == 1);
-    CHECK(actions.readString() == "connect('1.2.3.4', 80) -> 1");
-    CHECK(log.readString() == "");
+    CHECK(log.str() == "connect('1.2.3.4', 80) -> 1");
+    CHECK(output.str() == "");
   }
 
   SUBCASE("connect(const char*)") {
     int n = loggingClient.connect("1.2.3.4", 80);
 
     CHECK(n == 1);
-    CHECK(actions.readString() == "connect('1.2.3.4', 80) -> 1");
-    CHECK(log.readString() == "");
+    CHECK(log.str() == "connect('1.2.3.4', 80) -> 1");
+    CHECK(output.str() == "");
   }
 
   SUBCASE("connected()") {
     uint8_t n = loggingClient.connected();
 
     CHECK(n == false);
-    CHECK(actions.readString() == "connected() -> 0");
-    CHECK(log.readString() == "");
+    CHECK(log.str() == "connected() -> 0");
+    CHECK(output.str() == "");
   }
 
   SUBCASE("stop()") {
     loggingClient.stop();
 
-    CHECK(actions.readString() == "stop()");
-    CHECK(log.readString() == "");
+    CHECK(log.str() == "stop()");
+    CHECK(output.str() == "");
   }
 
   SUBCASE("operator bool()") {
     bool n = loggingClient.operator bool();
 
     CHECK(n == true);
-    CHECK(actions.readString() == "operator bool() -> true");
-    CHECK(log.readString() == "");
+    CHECK(log.str() == "operator bool() -> true");
+    CHECK(output.str() == "");
   }
 
   SUBCASE("peek()") {
@@ -76,8 +77,8 @@ TEST_CASE("ReadLoggingClient") {
     int n = loggingClient.peek();
 
     CHECK(n == 'A');
-    CHECK(actions.readString() == "peek() -> 65");
-    CHECK(log.readString() == "");
+    CHECK(log.str() == "peek() -> 65");
+    CHECK(output.str() == "");
   }
 
   SUBCASE("read()") {
@@ -86,8 +87,8 @@ TEST_CASE("ReadLoggingClient") {
     int n = loggingClient.read();
 
     CHECK(n == 'A');
-    CHECK(actions.readString() == "read() -> 65");
-    CHECK(log.readString() == "A");
+    CHECK(log.str() == "read() -> 65");
+    CHECK(output.str() == "A");
   }
 
   SUBCASE("read(uint8_t*,size_t)") {
@@ -97,8 +98,8 @@ TEST_CASE("ReadLoggingClient") {
     size_t n = loggingClient.read(s, 4);
 
     CHECK(n == 3);
-    CHECK(actions.readString() == "read(4) -> 3");
-    CHECK(log.readString() == "ABC");
+    CHECK(log.str() == "read(4) -> 3");
+    CHECK(output.str() == "ABC");
   }
 
   SUBCASE("readBytes()") {
@@ -108,9 +109,9 @@ TEST_CASE("ReadLoggingClient") {
     size_t n = loggingClient.readBytes(s, 4);
 
     CHECK(n == 3);
-    CHECK(log.readString() == "ABC");
+    CHECK(output.str() == "ABC");
 #if STREAMUTILS_STREAM_READBYTES_IS_VIRTUAL
-    CHECK(actions.readString() == "readBytes(4) -> 3");
+    CHECK(log.str() == "readBytes(4) -> 3");
 #endif
   }
 
@@ -118,21 +119,21 @@ TEST_CASE("ReadLoggingClient") {
     int n = loggingClient.write('A');
 
     CHECK(n == 1);
-    CHECK(actions.readString() == "write('A') -> 1");
-    CHECK(log.readString() == "");
+    CHECK(log.str() == "write('A') -> 1");
+    CHECK(output.str() == "");
   }
 
   SUBCASE("write(char*,size_t)") {
     int n = loggingClient.write("ABCDEF", 6);
 
     CHECK(n == 4);
-    CHECK(actions.readString() == "write('ABCDEF', 6) -> 4");
-    CHECK(log.readString() == "");
+    CHECK(log.str() == "write('ABCDEF', 6) -> 4");
+    CHECK(output.str() == "");
   }
 
   SUBCASE("flush()") {
     loggingClient.flush();
 
-    CHECK(actions.readString() == "flush()");
+    CHECK(log.str() == "flush()");
   }
 }