+++ extras/test/WriteBufferingClientTest.cpp
@@ -7,7 +7,7 @@
 #include "StreamUtils/Clients/MemoryClient.hpp"
 #include "StreamUtils/Clients/SpyingClient.hpp"
 #include "StreamUtils/Clients/WriteBufferingClient.hpp"
-#include "StreamUtils/Streams/MemoryStream.hpp"
+#include "StreamUtils/Prints/StringPrint.hpp"
 
 #include "doctest.h"
 
@@ -16,8 +16,8 @@ using namespace StreamUtils;
 TEST_CASE("WriteBufferingClient") {
   MemoryClient target(64);
 
-  MemoryStream actions(64);
-  SpyingClient spy{target, actions};
+  StringPrint log;
+  SpyingClient spy{target, log};
 
   GIVEN("capacity is 4") {
     WriteBufferingClient bufferingClient{spy, 4};
@@ -26,35 +26,35 @@ TEST_CASE("WriteBufferingClient") {
       target.print("ABC");
 
       CHECK(bufferingClient.available() == 3);
-      CHECK(actions.readString() == "available() -> 3");
+      CHECK(log.str() == "available() -> 3");
     }
 
     SUBCASE("connect(IPAddress)") {
       int n = bufferingClient.connect(IPAddress("1.2.3.4"), 80);
 
       CHECK(n == 1);
-      CHECK(actions.readString() == "connect('1.2.3.4', 80) -> 1");
+      CHECK(log.str() == "connect('1.2.3.4', 80) -> 1");
     }
 
     SUBCASE("connect(const char*)") {
       int n = bufferingClient.connect("1.2.3.4", 80);
 
       CHECK(n == 1);
-      CHECK(actions.readString() == "connect('1.2.3.4', 80) -> 1");
+      CHECK(log.str() == "connect('1.2.3.4', 80) -> 1");
     }
 
     SUBCASE("connected()") {
       uint8_t n = bufferingClient.connected();
 
       CHECK(n == false);
-      CHECK(actions.readString() == "connected() -> 0");
+      CHECK(log.str() == "connected() -> 0");
     }
 
     SUBCASE("stop()") {
       bufferingClient.write("ABC", 3);
       bufferingClient.stop();
 
-      CHECK(actions.readString() ==
+      CHECK(log.str() ==
             "write('ABC', 3) -> 3"
             "stop()");
     }
@@ -63,20 +63,20 @@ TEST_CASE("WriteBufferingClient") {
       bool n = bufferingClient.operator bool();
 
       CHECK(n == true);
-      CHECK(actions.readString() == "operator bool() -> true");
+      CHECK(log.str() == "operator bool() -> true");
     }
 
     SUBCASE("flush() forwards to target)") {
       bufferingClient.flush();
 
-      CHECK(actions.readString() == "flush()");
+      CHECK(log.str() == "flush()");
     }
 
     SUBCASE("flush() calls write() and flush()") {
       bufferingClient.write("ABC", 3);
       bufferingClient.flush();
 
-      CHECK(actions.readString() ==
+      CHECK(log.str() ==
             "write('ABC', 3) -> 3"
             "flush()");
     }
@@ -85,14 +85,14 @@ TEST_CASE("WriteBufferingClient") {
       target.print("ABC");
 
       CHECK(bufferingClient.peek() == 'A');
-      CHECK(actions.readString() == "peek() -> 65");
+      CHECK(log.str() == "peek() -> 65");
     }
 
     SUBCASE("read()") {
       target.print("ABC");
 
       CHECK(bufferingClient.read() == 'A');
-      CHECK(actions.readString() == "read() -> 65");
+      CHECK(log.str() == "read() -> 65");
     }
 
     SUBCASE("readBytes()") {
@@ -104,7 +104,7 @@ TEST_CASE("WriteBufferingClient") {
       CHECK(n == 3);
       CHECK(s == std::string("ABC"));
 #if STREAMUTILS_STREAM_READBYTES_IS_VIRTUAL
-      CHECK(actions.readString() == "readBytes(3) -> 3");
+      CHECK(log.str() == "readBytes(3) -> 3");
 #endif
     }
 
@@ -113,7 +113,7 @@ TEST_CASE("WriteBufferingClient") {
         int n = bufferingClient.write('A');
 
         CHECK(n == 1);
-        CHECK(actions.readString() == "");
+        CHECK(log.str() == "");
       }
 
       SUBCASE("write(uint8_t) should flush") {
@@ -123,27 +123,27 @@ TEST_CASE("WriteBufferingClient") {
         bufferingClient.write('D');
         bufferingClient.write('E');
 
-        CHECK(actions.readString() == "write('ABCD', 4) -> 4");
+        CHECK(log.str() == "write('ABCD', 4) -> 4");
       }
 
       SUBCASE("write(char*,3) goes in buffer") {
         size_t n = bufferingClient.write("ABC", 3);
 
         CHECK(n == 3);
-        CHECK(actions.readString() == "");
+        CHECK(log.str() == "");
       }
 
       SUBCASE("write(char*,4) bypasses buffer") {
         size_t n = bufferingClient.write("ABCD", 4);
 
         CHECK(n == 4);
-        CHECK(actions.readString() == "write('ABCD', 4) -> 4");
+        CHECK(log.str() == "write('ABCD', 4) -> 4");
       }
       SUBCASE("write(char*,2) bypasses buffer") {
         size_t n = bufferingClient.write("ABCD", 4);
 
         CHECK(n == 4);
-        CHECK(actions.readString() == "write('ABCD', 4) -> 4");
+        CHECK(log.str() == "write('ABCD', 4) -> 4");
       }
     }
 
@@ -154,14 +154,14 @@ TEST_CASE("WriteBufferingClient") {
         size_t n = bufferingClient.write("BCD", 3);
 
         CHECK(n == 3);
-        CHECK(actions.readString() == "write('ABCD', 4) -> 4");
+        CHECK(log.str() == "write('ABCD', 4) -> 4");
       }
 
       SUBCASE("write(char*,7) bypasses") {
         size_t n = bufferingClient.write("BCDEFGH", 7);
 
         CHECK(n == 7);
-        CHECK(actions.readString() ==
+        CHECK(log.str() ==
               "write('ABCD', 4) -> 4"
               "write('EFGH', 4) -> 4");
       }
@@ -175,20 +175,20 @@ TEST_CASE("WriteBufferingClient") {
       int n = bufferingClient.write('X');
 
       CHECK(n == 1);
-      CHECK(actions.readString() == "write('X') -> 1");
+      CHECK(log.str() == "write('X') -> 1");
     }
 
     SUBCASE("write(char*,1) forwards to target") {
       int n = bufferingClient.write("A", 1);
 
       CHECK(n == 1);
-      CHECK(actions.readString() == "write('A', 1) -> 1");
+      CHECK(log.str() == "write('A', 1) -> 1");
     }
 
     SUBCASE("flush() forwards to target") {
       bufferingClient.flush();
 
-      CHECK(actions.readString() == "flush()");
+      CHECK(log.str() == "flush()");
     }
   }
 
@@ -198,6 +198,6 @@ TEST_CASE("WriteBufferingClient") {
       bufferingClient.write("ABC", 3);
     }
 
-    CHECK(actions.readString() == "write('ABC', 3) -> 3");
+    CHECK(log.str() == "write('ABC', 3) -> 3");
   }
 }