+++ extras/test/WriteBufferingStreamTest.cpp
@@ -4,6 +4,7 @@
 
 #include "FailingAllocator.hpp"
 
+#include "StreamUtils/Prints/StringPrint.hpp"
 #include "StreamUtils/Streams/MemoryStream.hpp"
 #include "StreamUtils/Streams/SpyingStream.hpp"
 #include "StreamUtils/Streams/WriteBufferingStream.hpp"
@@ -15,8 +16,8 @@ using namespace StreamUtils;
 TEST_CASE("WriteBufferingStream") {
   MemoryStream upstream(64);
 
-  MemoryStream history(64);
-  SpyingStream spy{upstream, history};
+  StringPrint log;
+  SpyingStream spy{upstream, log};
 
   GIVEN("capacity is 4") {
     WriteBufferingStream stream{spy, 4};
@@ -25,20 +26,20 @@ TEST_CASE("WriteBufferingStream") {
       upstream.print("ABC");
 
       CHECK(stream.available() == 3);
-      CHECK(history.readString() == "available() -> 3");
+      CHECK(log.str() == "available() -> 3");
     }
 
     SUBCASE("flush() forwards to upstream)") {
       stream.flush();
 
-      CHECK(history.readString() == "flush()");
+      CHECK(log.str() == "flush()");
     }
 
     SUBCASE("flush() calls write() and flush()") {
       stream.write("ABC", 3);
       stream.flush();
 
-      CHECK(history.readString() ==
+      CHECK(log.str() ==
             "write('ABC', 3) -> 3"
             "flush()");
     }
@@ -47,14 +48,14 @@ TEST_CASE("WriteBufferingStream") {
       upstream.print("ABC");
 
       CHECK(stream.peek() == 'A');
-      CHECK(history.readString() == "peek() -> 65");
+      CHECK(log.str() == "peek() -> 65");
     }
 
     SUBCASE("read()") {
       upstream.print("ABC");
 
       CHECK(stream.read() == 'A');
-      CHECK(history.readString() == "read() -> 65");
+      CHECK(log.str() == "read() -> 65");
     }
 
     SUBCASE("readBytes()") {
@@ -66,7 +67,7 @@ TEST_CASE("WriteBufferingStream") {
       CHECK(n == 3);
       CHECK(s == std::string("ABC"));
 #if STREAMUTILS_STREAM_READBYTES_IS_VIRTUAL
-      CHECK(history.readString() == "readBytes(3) -> 3");
+      CHECK(log.str() == "readBytes(3) -> 3");
 #endif
     }
 
@@ -75,7 +76,7 @@ TEST_CASE("WriteBufferingStream") {
         int n = stream.write('A');
 
         CHECK(n == 1);
-        CHECK(history.readString() == "");
+        CHECK(log.str() == "");
       }
 
       SUBCASE("write(uint8_t) should flush") {
@@ -85,27 +86,27 @@ TEST_CASE("WriteBufferingStream") {
         stream.write('D');
         stream.write('E');
 
-        CHECK(history.readString() == "write('ABCD', 4) -> 4");
+        CHECK(log.str() == "write('ABCD', 4) -> 4");
       }
 
       SUBCASE("write(char*,3) goes in buffer") {
         size_t n = stream.write("ABC", 3);
 
         CHECK(n == 3);
-        CHECK(history.readString() == "");
+        CHECK(log.str() == "");
       }
 
       SUBCASE("write(char*,4) bypasses buffer") {
         size_t n = stream.write("ABCD", 4);
 
         CHECK(n == 4);
-        CHECK(history.readString() == "write('ABCD', 4) -> 4");
+        CHECK(log.str() == "write('ABCD', 4) -> 4");
       }
       SUBCASE("write(char*,2) bypasses buffer") {
         size_t n = stream.write("ABCD", 4);
 
         CHECK(n == 4);
-        CHECK(history.readString() == "write('ABCD', 4) -> 4");
+        CHECK(log.str() == "write('ABCD', 4) -> 4");
       }
     }
 
@@ -116,14 +117,14 @@ TEST_CASE("WriteBufferingStream") {
         size_t n = stream.write("BCD", 3);
 
         CHECK(n == 3);
-        CHECK(history.readString() == "write('ABCD', 4) -> 4");
+        CHECK(log.str() == "write('ABCD', 4) -> 4");
       }
 
       SUBCASE("write(char*,7) bypasses") {
         size_t n = stream.write("BCDEFGH", 7);
 
         CHECK(n == 7);
-        CHECK(history.readString() ==
+        CHECK(log.str() ==
               "write('ABCD', 4) -> 4"
               "write('EFGH', 4) -> 4");
       }
@@ -137,20 +138,20 @@ TEST_CASE("WriteBufferingStream") {
       int n = stream.write('X');
 
       CHECK(n == 1);
-      CHECK(history.readString() == "write('X') -> 1");
+      CHECK(log.str() == "write('X') -> 1");
     }
 
     SUBCASE("write(char*,1) forwards to upstream") {
       int n = stream.write("A", 1);
 
       CHECK(n == 1);
-      CHECK(history.readString() == "write('A', 1) -> 1");
+      CHECK(log.str() == "write('A', 1) -> 1");
     }
 
     SUBCASE("flush() forwards to upstream") {
       stream.flush();
 
-      CHECK(history.readString() == "flush()");
+      CHECK(log.str() == "flush()");
     }
   }
 
@@ -160,6 +161,6 @@ TEST_CASE("WriteBufferingStream") {
       stream.write("ABC", 3);
     }
 
-    CHECK(history.readString() == "write('ABC', 3) -> 3");
+    CHECK(log.str() == "write('ABC', 3) -> 3");
   }
 }