+++ extras/test/cores/avr/Print.h
@@ -6,9 +6,8 @@
 
 #include <stdint.h>
 #include <cstring>
-#include <string>
 
-using String = std::string;
+#include <WString.h>
 
 struct Print {
   virtual size_t write(const uint8_t *buffer, size_t size) = 0;