+++ extras/test/cores/avr/WString.h
@@ -1,0 +1,9 @@
+// StreamUtils - github.com/bblanchon/ArduinoStreamUtils
+// Copyright Benoit Blanchon 2019-2020
+// MIT License
+
+#pragma once
+
+#include <string>
+
+using String = std::string;