+++ extras/test/cores/samd/WString.h
@@ -1,0 +1,1 @@
+#include "../avr/WString.h"