+++ src/StreamUtils/Polyfills.hpp
@@ -1,0 +1,26 @@
+// StreamUtils - github.com/bblanchon/ArduinoStreamUtils
+// Copyright Benoit Blanchon 2019-2020
+// MIT License
+
+#pragma once
+
+namespace StreamUtils {
+namespace Polyfills {
+
+template <typename T>
+struct remove_reference {
+  using type = T;
+};
+
+template <typename T>
+struct remove_reference<T&> {
+  using type = T;
+};
+
+template <typename T>
+typename remove_reference<T>::type&& move(T&& t) {
+  return static_cast<typename remove_reference<T>::type&&>(t);
+}
+
+}  // namespace Polyfills
+}  // namespace StreamUtils